requirements.txt
# Django Core
Django>=5.0,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# WebSocket & Async
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0

# Computer Vision & ML
opencv-python-headless>=4.8.0
tensorflow>=2.15.0
Pillow>=10.0.0
numpy>=1.24.0

# Database
psycopg2-binary>=2.9.9

# Task Queue & Cache
celery>=5.3.0
redis>=5.0.0

# Cloud Storage
boto3>=1.34.0
django-storages>=1.14.0

# Production Server
gunicorn>=21.0.0
whitenoise>=6.6.0

# Development
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-django>=4.7.0
pytest-asyncio>=0.21.0

# Code Quality
black>=23.0.0
flake8>=6.1.0
isort>=5.12.0
